refactor(domain): split reconciliation.py into port + invariants (#237)
S4 of epic #233. domain/reconciliation.py held two concepts with different reasons to
change (SRP/screaming, C-200d). Split into:
- domain/reconciliation_port.py — Reconciler Protocol + RECONCILER_NOT_INJECTED_MSG (a
  stable abstraction, SAP).
- domain/reconciliation_invariants.py — ReconciliationInvariants (a concrete value object).

Updated all importers (ensemble.py, dataframe_ensemble.py, prediction_frame_ensemble.py,
reconcile_frames.py, adapter.py, modules/reconciliation/__init__.py, domain/__init__.py)
and renamed tests/test_domain/test_reconciliation.py → test_reconciliation_invariants.py.
No behavior change; old module deleted (no shim). Reconciliation/ensemble/domain tests green.
