+++ views_pipeline_core/domain/__init__.py
@@ -1,7 +1,7 @@
 from views_frames import SpatialLevel
 
 from .horizon import ForecastHorizon
-from .reconciliation import ReconciliationInvariants
+from .reconciliation_invariants import ReconciliationInvariants
 from .temporal import PartitionSet, TemporalPartition
 
 __all__ = [
