+++ views_pipeline_core/domain/reconciliation_port.py
@@ -1,0 +1,42 @@
+"""The reconciliation injection **port**: the `Reconciler` protocol (DIP).
+
+Ensemble managers depend on this abstraction, never on a concrete reconciliation package.
+The concrete, frames-native reconciler (`views_frames_reconcile.ReconciliationModule`, the
+views-frames v1.7.0 sibling) is injected at the composition root (views-models);
+pipeline-core never imports it (ADP — no cycle). Geography (the `(time, priogrid) -> country`
+mapping) is baked into the injected instance (views-frames ADR-014), so this port is frames-only.
+
+Split from the former `domain.reconciliation` (#237): the port is a *stable abstraction* (SAP)
+and changes for different reasons than the concrete `ReconciliationInvariants` value object,
+which now lives in `domain.reconciliation_invariants`.
+"""
+from typing import Protocol, runtime_checkable
+
+from views_frames import PredictionFrame
+
+
+@runtime_checkable
+class Reconciler(Protocol):
+    """Port for hierarchical pgm→cm forecast reconciliation (the DIP abstraction).
+
+    `reconcile` takes a country-level (cm) frame and a grid-level (pgm) frame and
+    returns a **new** pgm frame whose grid values are scaled so they sum to the cm
+    country totals. Frames-only by design: the geography mapping is held by the
+    concrete implementation, injected at the composition root.
+    """
+
+    def reconcile(
+        self, cm_frame: PredictionFrame, pgm_frame: PredictionFrame
+    ) -> PredictionFrame:
+        """Scale pgm forecasts to cm country totals; return a new pgm frame."""
+        ...
+
+
+#: Fail-loud message when reconciliation is configured but no concrete `Reconciler`
+#: was injected at the composition root. Single source of truth for both ensemble
+#: managers (no silent-off — see #194/#195).
+RECONCILER_NOT_INJECTED_MSG = (
+    "Reconciliation 'pgm_cm_point' is configured but no Reconciler was injected. "
+    "The composition root (the views-models ensemble main) must inject a concrete "
+    "Reconciler. See issue #195."
+)
